nhập hàng:

See merge request dev-teca-qlnv/qlnv-hang!41

+++ src/main/java/com/tcdt/qlnvhang/request/search/HhQdNhapxuatSearchReq.java
@@ -27,7 +27,7 @@ public class HhQdNhapxuatSearchReq extends BaseRequest {
 	String maDvi;
 	String maVthh;
 	String loaiQd;
-	Integer namNhap;
+	Long namNhap;
 	String veViec;
 	String loaiVthh;
 	String trichYeu;
